+++ app/src/main/java/com/rama/mako/activities/settings/SettingsCheckboxController.kt
@@ -75,6 +75,8 @@ class SettingsCheckboxController(private val activity: SettingsActivity) {
         )
         bindWdCheckbox(R.id.show_profile_indicator, FileKeys.APPS_PROFILE_INDICATOR, true)
         bindWdCheckbox(R.id.multi_column, FileKeys.APPS_MULTI_COLUMN, false)
+        bindWdCheckbox(R.id.show_api_indicators, FileKeys.APPS_SHOW_API_INDICATORS, false)
+        bindWdCheckbox(R.id.show_app_size, FileKeys.APPS_SHOW_SIZE, false)
 
         bindWdCheckbox(
             R.id.lock_settings,