+++ app/src/main/java/com/rama/mako/managers/PrefsManager.kt
@@ -24,6 +24,8 @@ class PrefsManager private constructor(context: Context) : BohioPrefsManager(con
         const val APPS_ICON_PACK_PACKAGE = "apps:icon_pack_package"
         const val APPS_ICONS_OPEN_SETTINGS = "apps:icons:open_settings"
         const val APPS_MULTI_COLUMN = "apps:multi_column"
+        const val APPS_SHOW_API_INDICATORS = "apps:show_api_indicators"
+        const val APPS_SHOW_SIZE = "apps:show_size"
 
         const val HOME_BACKGROUND_MODE = "home:background_mode"
         const val HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH =
@@ -143,6 +145,8 @@ class PrefsManager private constructor(context: Context) : BohioPrefsManager(con
         editor.putString(FileKeys.APPS_ICON_PACK_PACKAGE, "")
         editor.putBoolean(FileKeys.APPS_ICONS_OPEN_SETTINGS, true)
         editor.putBoolean(FileKeys.APPS_MULTI_COLUMN, false)
+        editor.putBoolean(FileKeys.APPS_SHOW_API_INDICATORS, false)
+        editor.putBoolean(FileKeys.APPS_SHOW_SIZE, false)
 
         editor.putString(FileKeys.HOME_BACKGROUND_MODE, BackgroundMode.DEFAULT)
         editor.putInt(FileKeys.HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH, 9)
@@ -379,6 +383,12 @@ class PrefsManager private constructor(context: Context) : BohioPrefsManager(con
 
     fun isMultiColumnEnabled(): Boolean =
         prefs.getBoolean(FileKeys.APPS_MULTI_COLUMN, false)
+
+    fun hasApiIndicatorsVisible(): Boolean =
+        prefs.getBoolean(FileKeys.APPS_SHOW_API_INDICATORS, false)
+
+    fun hasAppSizeVisible(): Boolean =
+        prefs.getBoolean(FileKeys.APPS_SHOW_SIZE, false)
 
     fun setMultiColumnEnabled(enabled: Boolean) =
         prefs.edit().putBoolean(FileKeys.APPS_MULTI_COLUMN, enabled).apply()